Create a shared it.json - work in progress
Dear All I started to create the italian version with my team in @identitadigitaledefender group. We are tech people and legal staff and we are "GDPR open source" oriented. I uploaded here the file partially completed. I think this translaltion, for be user friendly and alligned with italian legal/tech terms (in the specific context) should be done online with other italian delegate in the team. 
So, I took the task to start the work and I shared the file here for give freedom to other italian partecipant to update it directly in the common github workspace. I'll continuous to work on it and any external support is welcome. Thanks to French colleagues (and all the others) which started this very nice project! Good work to all and merry christmas.